Move DragContainer class
