+++ src/amulet_editor/models/widgets/__init__.py
@@ -1,5 +1,4 @@
 # from ._card import ALinkCard, QPixCard
-from ._draggable import ADragContainer
 from ._icon import AStylableSvgWidget, AIconButton, ATooltipIconButton
 
 # from ._label import QElidedLabel
